Move simulation menu to simulation page
