+++ components/simulation-page/simulation-menu/legende/Legende.module.scss
