+++ components/simulation-page/simulation-menu/legende/Legende.tsx
